+++ trunk/mov/src/org/mov/quote/InsightTraderQuoteFilter.java
@@ -89,13 +89,13 @@ public class InsightTraderQuoteFilter implements EODQuoteFilter {
 
 		// Convert all prices from cents to dollars
                 try {
-                    float day_open = Float.parseFloat(quoteParts[i++]) / 100;
-                    float day_high = Float.parseFloat(quoteParts[i++]) / 100;
-                    float day_low = Float.parseFloat(quoteParts[i++]) / 100;
-                    float day_close = Float.parseFloat(quoteParts[i++]) / 100;
+                    double day_open = Double.parseDouble(quoteParts[i++]) / 100.0;
+                    double day_high = Double.parseDouble(quoteParts[i++]) / 100.0;
+                    double day_low = Double.parseDouble(quoteParts[i++]) / 100.0;
+                    double day_close = Double.parseDouble(quoteParts[i++]) / 100.0;
                     
                     // Convert volume from 1/100th volume to real volume
-                    int day_volume = Integer.parseInt(quoteParts[i++]) * 100;
+                    long day_volume = Long.parseLong(quoteParts[i++]) * 100;
                     quote = new EODQuote(symbol, date, day_volume, day_low, day_high,
                                          day_open, day_close);
                 }
